+++ src/bpm/commands/root.go
@@ -216,8 +216,9 @@ func isRunningSystemd() bool {
 }
 
 // Occasionally RunC can get in an inconsistent state after a restart where
-// it's internal state.json file is truncated. RunC is unable to get out of
-// this state without some intervention. This is that intervention.
+// it's internal state.json file is truncated or missing but the parent
+// directory still exists. RunC is unable to get out of this state without some
+// intervention. This is that intervention.
 func forceCleanupBrokenRuncState(logger lager.Logger, runcLifecycle *lifecycle.RuncLifecycle) error {
 	// We compute this here rather than adding a new function to the
 	// configuration object to try and contain this hack to one place.